+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 -   Unify Silicon for IF and EnderIO [\#131](https://github.com/EnigmaticaModpacks/Enigmatica10/pull/131)
 -   Cheaper Source Blocks! [\#131](https://github.com/EnigmaticaModpacks/Enigmatica10/pull/131)
 -   Fluix may be crafted with Industrial Foregoing [\#131](https://github.com/EnigmaticaModpacks/Enigmatica10/pull/131)
+-   Leads and Sticky Pistons now accept pink slime [\#133](https://github.com/EnigmaticaModpacks/Enigmatica10/pull/133)
 
 #### 🦟 Bugs Fixed